WIP: add test
